Update README.md
Here the 2 players are named as player left and player right, who each control the beams/ bat that hit the ball, which randomly bounce as it hit the top, bottom and the slabs. 
If the ball crosses the left or right side of the screen, the opposite side gets a point.
After a certain number of points(currently 10) are scored, The winner is declared by turning the top fount name to green and the loser's to red. A Dialog box is also shows the winner.